LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/semantics/popovers/popover-focus-expected.txt

This is a popover

autofocus button  second autofocus button

FAIL Popover focus test: default behavior - popover is not focused assert_equals: default behavior - popover is not focused activated by popover.showPopover() expected Element node <button id="priorFocus"></button> but got Element node <button>first button</button>
FAIL Popover button click focus test: default behavior - popover is not focused assert_false: popover should start out hidden expected false got true
PASS Popover corner cases test: default behavior - popover is not focused
FAIL Popover focus test: autofocus popover assert_equals: autofocus popover activated by popover.showPopover() expected Element node <div popover="" data-test="autofocus popover" autofocus="... but got Element node <button id="priorFocus"></button>
FAIL Popover button click focus test: autofocus popover assert_false: popover should start out hidden expected false got true
FAIL Popover corner cases test: autofocus popover assert_false: popover should start out hidden expected false got true
FAIL Popover focus test: autofocus empty popover assert_equals: autofocus empty popover activated by popover.showPopover() expected Element node <div popover="" data-test="autofocus empty popover" autof... but got Element node <button id="priorFocus"></button>
FAIL Popover button click focus test: autofocus empty popover assert_false: popover should start out hidden expected false got true
FAIL Popover corner cases test: autofocus empty popover assert_false: popover should start out hidden expected false got true
FAIL Popover focus test: autofocus popover with button assert_equals: autofocus popover with button activated by popover.showPopover() expected Element node <div popover="" data-test="autofocus popover with button"... but got Element node <button>button</button>
FAIL Popover button click focus test: autofocus popover with button assert_false: popover should start out hidden expected false got true
FAIL Popover corner cases test: autofocus popover with button assert_false: popover should start out hidden expected false got true
FAIL Popover focus test: autofocus child assert_equals: prior element should get focus after Escape expected Element node <button id="priorFocus"></button> but got Element node <button autofocus="" class="should-be-focused">autofocus ...
FAIL Popover button click focus test: autofocus child assert_false: popover should start out hidden expected false got true
FAIL Popover corner cases test: autofocus child assert_false: popover should start out hidden expected false got true
FAIL Popover focus test: autofocus on tabindex=0 element assert_equals: prior element should get focus after Escape expected Element node <button id="priorFocus"></button> but got Element node <p autofocus="" tabindex="0" class="should-be-focused">Th...
FAIL Popover button click focus test: autofocus on tabindex=0 element assert_false: popover should start out hidden expected false got true
FAIL Popover corner cases test: autofocus on tabindex=0 element assert_false: popover should start out hidden expected false got true
FAIL Popover focus test: autofocus multiple children assert_equals: prior element should get focus after Escape expected Element node <button id="priorFocus"></button> but got Element node <button autofocus="" class="should-be-focused">autofocus ...
FAIL Popover button click focus test: autofocus multiple children assert_false: popover should start out hidden expected false got true
FAIL Popover corner cases test: autofocus multiple children assert_false: popover should start out hidden expected false got true
FAIL Popover focus test: autofocus popover and multiple autofocus children assert_equals: autofocus popover and multiple autofocus children activated by popover.showPopover() expected Element node <div popover="" autofocus="" tabindex="-1" data-test="aut... but got Element node <button autofocus="">autofocus button</button>
FAIL Popover button click focus test: autofocus popover and multiple autofocus children assert_false: popover should start out hidden expected false got true
FAIL Popover corner cases test: autofocus popover and multiple autofocus children assert_false: popover should start out hidden expected false got true
